backend/requirements.txt
# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Vector DB
chromadb>=0.4.22

# Embeddings (runs locally, no API cost)
sentence-transformers>=2.6.0

# LLM — Gemini
google-generativeai>=0.5.0

# PDF parsing
PyMuPDF>=1.23.0

# URL scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.27.0

# YouTube transcripts
youtube-transcript-api>=0.6.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
aiofiles>=23.2.0
